Add consort_bulkload.pl version from 2012
consort_bulkload.pl is the script that uploads each of the consortium flat text data files.  It tests the data for completeness and corrects errors.  It produces log files for the changes and the errors.  It then created a series of ".in" text files of the separate text files that will eventually become the database hash files.
This file is the original 2012 version of this script, created by Dick Moe.  This is the earliest copy I could find.  This file was uploaded first in order to retroactively document changes to this file that have been done by David Baxter and now by me.
The copy of this file that is being updates on the BIM computer in the Herbarium is in the directory:
\CCH\bulkload_data\consort_bulkload.pl